requirements.txt
Django==1.9.7
Pillow==3.2.0
argparse==1.4.0
beautifulsoup4==4.4.1
boto==2.41.0
click==6.6
coverage==4.1
dj-database-url==0.4.1
django-appconf==1.0.2
django-classy-tags==0.7.2
django-compressor==2.0
django-filter==0.13.0
django-htmlmin==0.9.1
django-mailgun==0.9.1
django-sekizai==0.9.0
django-storages==1.4.1
django-tenants==1.1.7
djangorestframework==3.3.3
gunicorn==19.6.0
html5lib==0.9999999
ordereddict==1.1
psycopg2==2.6.1
python-dotenv==0.5.1
rcssmin==1.0.6
requests==2.10.0
rjsmin==1.0.12
six==1.10.0
